Moved many classes to their own files for organization purposes and deleted unused files and classes.
