--- lib/src/entities/metadata_extractor.dart
@@ -1,11 +1,0 @@
-class Metadata {
-  final String title;
-  final String author;
-  final String year;
-
-  Metadata(
-    this.title,
-    this.author,
-    this.year,
-  );
-}
